TIN-92: Add filters, sort, visible, change width and save on db
TIN-92: fix test and fixture

TIN-92: add alembic

changes from prod

TIN-92: Updated CI/CD

TIN-92: fix sort all tests; fix sort in routes for relationship table

+++ web_app/routes/form_of_ownerships.py
@@ -1,8 +1,12 @@
-from fastapi import HTTPException, APIRouter, Depends, status
-from typing import List
+from fastapi import HTTPException, APIRouter, Depends, status, Query
+from typing import Annotated, Optional, List
+
+from sqlalchemy import and_, or_
 from sqlalchemy.ext.asyncio import AsyncSession
 from web_app.database import get_db
 from sqlalchemy.future import select
+
+from web_app.models import Customers
 from web_app.models.form_of_ownerships import FormOfOwnerships
 from web_app.schemas.form_of_ownership import (
     FormOfOwnershipCreate,
@@ -15,14 +19,42 @@ router = APIRouter()
 
 
 # Endpoints
-@router.get("/form-of-ownership", response_model=List[FormOfOwnershipResponse])
+
+
+@router.get("/form-of-ownership", response_model=list[FormOfOwnershipResponse])
 async def get_form_of_ownerships(
     db: AsyncSession = Depends(get_db),
+    id: Annotated[list[int] | None, Query()] = None,
+    name: Annotated[list[str] | None, Query()] = None,
     user_data: dict = Depends(token_verification_dependency),
+    sortBy: Optional[str] = None,
+    sortDir: Optional[str] = "asc",
 ):
     stmt = select(FormOfOwnerships)
+
+    filters = []
+
+    if id:
+        filters.append((FormOfOwnerships.id.in_(id)))
+    if name:
+        filters.append(or_(FormOfOwnerships.name.ilike(f"%{n}%") for n in name))
+
+    if filters:
+        stmt = stmt.where(and_(*filters))
+
+    if sortBy:
+        try:
+            sort_column = getattr(FormOfOwnerships, sortBy)
+            if sortDir.lower() == "desc":
+                stmt = stmt.order_by(sort_column.desc())
+            else:
+                stmt = stmt.order_by(sort_column.asc())
+        except AttributeError:
+            raise HTTPException(status_code=404, detail="Поле не найдено")
+
     result = await db.execute(stmt)
     form_of_ownerships = result.scalars().all()
+
     return form_of_ownerships
 
 
@@ -100,6 +132,15 @@ async def delete_form_of_ownership(
     form_of_ownership = result.scalar_one_or_none()
     if not form_of_ownership:
         raise HTTPException(status_code=404, detail="Форма собственности не найдена")
+
+    customers_exist = await db.execute(
+        select(Customers).filter(Customers.form == object_id).limit(1)
+    )
+    if customers_exist.scalar_one_or_none():
+        raise HTTPException(
+            status_code=400,
+            detail="Невозможно удалить форму собственности: существуют связанные заказчики. Удалите их сначала.",
+        )
 
     # Удаление объекта
     await db.delete(form_of_ownership)
